+++ integrations/storage/v2/storage.go
@@ -31,40 +31,54 @@ import (
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
+	pbImplAuthorizationV1Shared "github.com/arangodb/kube-arangodb/integrations/authorization/v1/shared"
 	pbStorageV2 "github.com/arangodb/kube-arangodb/integrations/storage/v2/definition"
 	pbImplStorageV2Shared "github.com/arangodb/kube-arangodb/integrations/storage/v2/shared"
 	"github.com/arangodb/kube-arangodb/pkg/util"
+	utilConstantsContext "github.com/arangodb/kube-arangodb/pkg/util/constants/context"
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
-	"github.com/arangodb/kube-arangodb/pkg/util/shutdown"
 	"github.com/arangodb/kube-arangodb/pkg/util/svc"
+	"github.com/arangodb/kube-arangodb/pkg/util/svc/authenticator"
 )
 
 var _ pbStorageV2.StorageV2Server = &implementation{}
 var _ svc.Handler = &implementation{}
 
-func New(cfg Configuration) (svc.Handler, error) {
-	return newInternal(cfg)
+func New(ctx context.Context, cfg Configuration) (svc.Handler, error) {
+	return newInternal(ctx, cfg)
 }
 
-func newInternal(c Configuration) (*implementation, error) {
+func newInternal(ctx context.Context, c Configuration) (*implementation, error) {
 	if err := c.Validate(); err != nil {
 		return nil, errors.Wrapf(err, "Invalid config")
 	}
 
-	io, err := c.IO(shutdown.Context())
+	auth, ok := utilConstantsContext.AuthZClientPlugin.Get(ctx)
+	if !ok {
+		return nil, errors.Errorf("Unable to get AuthZ Client Plugin")
+	}
+
+	io, err := c.IO(ctx)
 	if err != nil {
 		return nil, err
 	}
 
 	return &implementation{
-		io: io,
+		io:   io,
+		auth: auth,
 	}, nil
 }
 
 type implementation struct {
 	io pbImplStorageV2Shared.IO
 
+	auth pbImplAuthorizationV1Shared.Evaluator
+
 	pbStorageV2.UnimplementedStorageV2Server
+}
+
+func (i *implementation) checkPermission(ctx context.Context, action, resource string) error {
+	return authenticator.GetIdentity(ctx).EvaluatePermission(ctx, i.auth, action, resource)
 }
 
 func (i *implementation) Name() string {
@@ -98,6 +112,10 @@ func (i *implementation) WriteObject(server pbStorageV2.StorageV2_WriteObjectSer
 	if path == "" {
 		log.Debug("path missing")
 		return status.Error(codes.InvalidArgument, "path missing")
+	}
+
+	if err := i.checkPermission(ctx, "storage:WriteObject", path); err != nil {
+		return err
 	}
 
 	wd, err := i.io.Write(ctx, path)
@@ -158,6 +176,10 @@ func (i *implementation) ReadObject(req *pbStorageV2.StorageV2ReadObjectRequest,
 		return status.Errorf(codes.InvalidArgument, "path missing")
 	}
 
+	if err := i.checkPermission(ctx, "storage:ReadObject", path); err != nil {
+		return err
+	}
+
 	rd, err := i.io.Read(ctx, path)
 	if err != nil {
 		if errors.Is(err, os.ErrNotExist) {
@@ -202,6 +224,10 @@ func (i *implementation) HeadObject(ctx context.Context, req *pbStorageV2.Storag
 		return nil, status.Error(codes.InvalidArgument, "path missing")
 	}
 
+	if err := i.checkPermission(ctx, "storage:HeadObject", path); err != nil {
+		return nil, err
+	}
+
 	info, err := i.io.Head(ctx, path)
 	if err != nil {
 		log.Err(err).Debug("getObjectInfo failed")
@@ -228,6 +254,10 @@ func (i *implementation) DeleteObject(ctx context.Context, req *pbStorageV2.Stor
 		return nil, status.Error(codes.InvalidArgument, "path missing")
 	}
 
+	if err := i.checkPermission(ctx, "storage:DeleteObject", path); err != nil {
+		return nil, err
+	}
+
 	deleted, err := i.io.Delete(ctx, path)
 	if err != nil {
 		log.Err(err).Debug("deleteObject failed")
@@ -245,6 +275,10 @@ func (i *implementation) ListObjects(req *pbStorageV2.StorageV2ListObjectsReques
 	log := logger.Str("func", "ReadObject").Str("path", req.GetPath().GetPath())
 	ctx := server.Context()
 	path := req.GetPath().GetPath()
+
+	if err := i.checkPermission(ctx, "storage:ListObjects", path); err != nil {
+		return err
+	}
 
 	lister, err := i.io.List(ctx, path)
 	if err != nil {
@@ -286,6 +320,10 @@ func (i *implementation) ListObjects(req *pbStorageV2.StorageV2ListObjectsReques
 }
 
 func (i *implementation) Init(ctx context.Context, in *pbStorageV2.StorageV2InitRequest) (*pbStorageV2.StorageV2InitResponse, error) {
+	if err := i.checkPermission(ctx, "storage:Init", ""); err != nil {
+		return nil, err
+	}
+
 	if err := i.io.Init(ctx, &pbImplStorageV2Shared.InitOptions{
 		Create: util.NewPointer(in.Create),
 	}); err != nil {