+++ pkg/apis/deployment/v1/conditions.go
@@ -137,6 +137,9 @@ const (
 
 	// ConditionTypeProfilesReady indicates that the profiles are ready.
 	ConditionTypeProfilesReady ConditionType = "ProfilesReady"
+
+	// ConditionTypeStorageReady indicates that the storage is ready.
+	ConditionTypeStorageReady ConditionType = "StorageReady"
 )
 
 // Condition represents one current condition of a deployment or deployment member.