+++ pkg/integrations/storage_v2.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -88,7 +88,7 @@ func (b *storageV2) Register(cmd *cobra.Command, fs FlagEnvHandler) error {
 }
 
 func (b *storageV2) Handler(ctx context.Context, cmd *cobra.Command) (svc.Handler, error) {
-	return pbImplStorageV2.New(b.Configuration)
+	return pbImplStorageV2.New(ctx, b.Configuration)
 }
 
 func (*storageV2) Init(ctx context.Context, cmd *cobra.Command) error {